Init and remove contact state actions factored out
