+++ plugin-hrm-form/src/states/ActionTypes.js
@@ -2,8 +2,6 @@ export const HANDLE_BLUR = 'HANDLE_BLUR';
 export const HANDLE_VALIDATE_FORM = 'HANDLE_VALIDATE_FORM';
 export const HANDLE_FOCUS = 'HANDLE_FOCUS';
 export const HANDLE_CHANGE = 'HANDLE_CHANGE';
-export const INITIALIZE_CONTACT_STATE = 'INITIALIZE_CONTACT_STATE';
-export const REMOVE_CONTACT_STATE = 'REMOVE_CONTACT_STATE';
 export const SAVE_END_MILLIS = 'SAVE_END_MILLIS';
 export const CHANGE_TAB = 'CHANGE_TAB';
 export const RESTORE_ENTIRE_FORM = 'RESTORE_ENTIRE_FORM';
