+++ plugin-hrm-form/src/utils/setUpActions.js
@@ -7,6 +7,7 @@ import { saveInsightsData } from '../services/InsightsService';
 import { transferChatStart } from '../services/ServerlessService';
 import { namespace, contactFormsBase } from '../states';
 import { Actions } from '../states/ContactState';
+import * as GeneralActions from '../states/actions';
 import { channelTypes, transferModes } from '../states/DomainConstants';
 import * as TransferHelpers from './transfer';
 import { saveFormSharedState, loadFormSharedState } from './sharedState';
@@ -43,7 +44,7 @@ const fromActionFunction = fun => async (payload, original) => {
  * @param {{ task: any }} payload
  */
 export const initializeContactForm = payload => {
-  Manager.getInstance().store.dispatch(Actions.initializeContactState(payload.task.taskSid));
+  Manager.getInstance().store.dispatch(GeneralActions.initializeContactState(payload.task.taskSid));
 };
 
 /**
@@ -203,5 +204,5 @@ export const sendInsightsData = setupObject => async payload => {
  */
 export const removeContactForm = payload => {
   const manager = Manager.getInstance();
-  manager.store.dispatch(Actions.removeContactState(payload.task.taskSid));
+  manager.store.dispatch(GeneralActions.removeContactState(payload.task.taskSid));
 };
